Не работает отображение единиц измерения

+++ sources/Device/src/Nextion/Controls/Label.cpp
@@ -6,7 +6,7 @@
 
 
 Label::Label(int _x, int _y, int _w, int _h, int _font, pchar _textRU, pchar _textEN, const Color &_colorText, bool _h_align, const Color &_colorBack) :
-    x(_x), y(_y), width(_w), height(_h), font(_font), h_aligned(_h_align), colorText(_colorText), colorBack(_colorBack)
+    x((int16)_x), y((int16)_y), width((int16)_w), height((int16)_h), font(_font), h_aligned(_h_align), colorText(_colorText), colorBack(_colorBack)
 {
     std::strcpy(text[0], _textRU);
     std::strcpy(text[1], _textEN);
